requirements.txt
# Core dependencies
pydantic>=2.0.0
requests>=2.28.0
PyYAML>=6.0

# Web dependencies (optional)
fastapi>=0.95.0
uvicorn>=0.22.0
jinja2>=3.1.2

# Database
SQLite3>=0.0.0  # Built into Python, but listed for clarity